# docs/architecture/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: UnFranchise Marketing App API
  description: |
    API for the UnFranchise Marketing App - A corporate-controlled content sharing
    and engagement platform for UnFranchise Owners (UFOs).

    ## Authentication
    Most endpoints require Bearer token authentication via JWT.
    Include the access token in the Authorization header:
    ```
    Authorization: Bearer {access_token}
    ```

    ## Rate Limiting
    - Authentication endpoints: 5 requests per 15 minutes
    - Standard API endpoints: 100 requests per minute
    - Share endpoints: 500 requests per hour
    - Upload endpoints: 20 requests per hour

    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Pagination
    List endpoints support pagination via query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 20, max: 100)
    - `sort`: Sort field with optional `-` prefix for descending (e.g., `-createdAt`)

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@unfranchise.com
  license:
    name: Proprietary
servers:
  - url: https://api.unfranchise.com/v1
    description: Production server
  - url: https://staging-api.unfranchise.com/v1
...
paths:
...
  /contacts:
    post:
      tags: [Contacts]
      summary: Create contact
      description: Add new contact to user's contact list
      operationId: createContact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '201':
          description: Contact created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Contact'
                  meta:
                    $ref: '#/components/schemas/Meta'

    get:
      tags: [Contacts]
      summary: List contacts
      description: Get user's contacts with filtering and search
      operationId: listContacts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema:
            type: string
          description: Search in name, email, company
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated tag names
        - name: relationshipType
          in: query
          schema:
            type: string
            enum: [prospect, customer, distributor, partner, other]
      responses:
        '200':
          description: Contacts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /contacts/{contactId}:
    get:
      tags: [Contacts]
      summary: Get contact details
      description: Get detailed contact information
      operationId: getContact
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Contact retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ContactDetail'

    put:
      tags: [Contacts]
      summary: Update contact
      description: Update contact information
      operationId: updateContact
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactUpdate'
      responses:
        '200':
          description: Contact updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Contact'

    delete:
      tags: [Contacts]
      summary: Delete contact
      description: Delete contact from user's list
      operationId: deleteContact
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Contact deleted successfully
  /contacts/import:
    post:
      tags: [Contacts]
      summary: Import contacts
      description: Import contacts from CSV file
      operationId: importContacts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with contacts
      responses:
        '202':
          description: Import job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      jobId:
                        type: string
                        format: uuid
                      status:
                        type: string
                        example: processing
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...
    ContactCreate:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        mobile:
          type: string
        company:
          type: string
        jobTitle:
          type: string
        relationshipType:
          type: string
          enum: [prospect, customer, distributor, partner, other]
        tags:
          type: array
          items:
            type: string
        source:
          type: string
        consentEmail:
          type: boolean
        consentSms:
          type: boolean
        notes:
          type: string
    ContactUpdate:
      allOf:
        - $ref: '#/components/schemas/ContactCreate'
    Contact:
      allOf:
        - $ref: '#/components/schemas/ContactCreate'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            engagementScore:
              type: integer
            lastEngagementDate:
              type: string
              format: date-time
            totalSharesReceived:
              type: integer
            totalClicks:
              type: integer
            createdAt:
              type: string
              format: date-time
    ContactDetail:
      allOf:
        - $ref: '#/components/schemas/Contact'
        - type: object
          properties:
            recentEngagements:
              type: array
              items:
                type: object
                properties:
                  eventType:
                    type: string
                  contentTitle:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
...